updated delete with exceptions

+++ src/main/java/com/be/elamreader/Controllers/BookController.java
@@ -5,6 +5,7 @@ import com.be.elamreader.Repositories.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -54,8 +55,17 @@ public class BookController {
     }
 
     @DeleteMapping("/{bookId}")
+    @Transactional
     protected ResponseEntity<?> deleteBookbyBookId(@PathVariable UUID bookId){
-        Book deletedBook = bookRepository.deleteBookByBookId(bookId);
-        return ResponseEntity.status(HttpStatus.ACCEPTED).body("book deleted" + deletedBook);
+        try {
+            if (bookRepository.getBookByBookId(bookId) == null) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("book not found");
+            }
+            bookRepository.deleteBookByBookId(bookId);
+            return ResponseEntity.status(HttpStatus.ACCEPTED).body("book deleted!");
+        }
+        catch (Exception e){
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("error deleting book "+e.getMessage());
+        }
     }
 }
